Adding JWT token authentication and authorization

+++ src/CostaRicaApi/Controllers/ExpensesController.cs
@@ -4,44 +4,66 @@ using CostaRicaApi.Models;
 using CostaRicaApi.Repositories;
 using CostaRicaApi.Dtos;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
+using System.Linq;
+using System.Security.Claims;
+using Microsoft.Extensions.Logging;
 
 namespace CostaRicaApi.Controllers {
 
+    [Authorize]
     [Route("api/v1/[controller]")]
     [ApiController]
     public class ExpensesController : ControllerBase {
 
         private readonly IExpenseRepository _expenseRepo;
+        private readonly IUserRepository _userRepo;
         private readonly IMapper _mapper;
-        public ExpensesController(IExpenseRepository expenseRepo, IMapper mapper)
+        private readonly ILogger<ExpensesController> _logger;
+        public ExpensesController(IExpenseRepository expenseRepo, IMapper mapper, IUserRepository userRepo, ILogger<ExpensesController> logger)
         {
             _expenseRepo = expenseRepo;
+            _userRepo = userRepo;
             _mapper = mapper;
+            _logger = logger;
         }
 
         [HttpGet]
-        public async Task<ActionResult<List<Expense>>> GetExpenses() {
-            var expenses = await _expenseRepo.GetAllExpensesAsync();
+        public async Task<ActionResult<List<ShallowExpenseDto>>> GetExpenses() {
+            var currentUserId = int.Parse(User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier).Value);
 
-            return expenses;
+            var expenses = await _expenseRepo.GetAllExpensesAsync(currentUserId);
+
+            return expenses.Select(x => _mapper.Map<ShallowExpenseDto>(x)).ToList();
         }
 
         [HttpGet("{id}")]
-        public async Task<ActionResult<Expense>> GetExpense(int id) {
-            var expense = await _expenseRepo.GetExpenseByIdAsync(id);
+        public async Task<ActionResult<ShallowExpenseDto>> GetExpense(int id) {
+            var currentUserId = int.Parse(User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier).Value);
 
+            var expense = await _expenseRepo.GetExpenseByIdAsync(id);
+            
             if(expense == null) {
                 return NotFound();
             }
 
-            return Ok(expense);
+            var expenseDto = _mapper.Map<ShallowExpenseDto>(expense);
+
+            if(expenseDto.OwnerId != currentUserId) {
+                return Unauthorized();
+            }
+
+            return Ok(expenseDto);
         }
 
         [HttpPost]
         public async Task<ActionResult<Expense>> AddExpense(ExpenseCreateUpdateDto dto) {
+            var currentUserId = int.Parse(User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier).Value);
+
             var expense = _mapper.Map<Expense>(dto);
+            expense.Owner = await _userRepo.GetUserByIdAsync(currentUserId);
+
             expense = await _expenseRepo.AddExpenseAsync(expense);
 
             if(0 == await _expenseRepo.SaveChangesAsync()) {
@@ -53,7 +75,12 @@ namespace CostaRicaApi.Controllers {
 
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateExpense(int id, ExpenseCreateUpdateDto dto) {
+            var currentUserId = int.Parse(User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier).Value);
             var expense = await _expenseRepo.GetExpenseByIdAsync(id);
+
+            if(expense.Owner.Id != currentUserId) {
+                return Unauthorized();
+            }
 
             if(expense == null) {
                 return NotFound();
@@ -70,10 +97,16 @@ namespace CostaRicaApi.Controllers {
 
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteExpense(int id) {
+            var currentUserId = int.Parse(User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier).Value);
+
             var expense = await _expenseRepo.GetExpenseByIdAsync(id);
 
             if(expense == null) {
                 return NotFound();
+            }
+
+            if(expense.Owner.Id != currentUserId) {
+                return Unauthorized();
             }
 
             _expenseRepo.RemoveExpense(expense);
